+++ liquid2/lexer.py
@@ -30,6 +30,7 @@ from .token import WhitespaceControl
 from .token import is_token_type
 
 if TYPE_CHECKING:
+    from .environment import Environment
     from .token import TokenT
 
 
@@ -178,6 +179,7 @@ class Lexer:
     TOKEN_RULES = _compile(NUMBERS, SYMBOLS, WORD)
 
     __slots__ = (
+        "env",
         "in_range",
         "line_start",
         "line_statements",
@@ -194,7 +196,9 @@ class Lexer:
         "template_string_stack",
     )
 
-    def __init__(self, source: str) -> None:
+    def __init__(self, env: Environment, source: str) -> None:
+        self.env = env
+
         self.markup: list[TokenT] = []
         """Markup resulting from scanning a Liquid template."""
 
@@ -314,8 +318,15 @@ class Lexer:
                     self.pos += match.end() - match.start()
                     self.start = self.pos
                     self.path_stack[-1].stop = self.pos
+                elif self.env.shorthand_indexes:
+                    if match := self.RE_INDEX.match(self.source, self.pos):
+                        self.path_stack[-1].path.append(int(match.group()))
+                        self.pos += match.end() - match.start()
+                        self.start = self.pos
+                    else:
+                        self.error("array indexes must use bracket notation")
                 else:
-                    self.error("expected a property name")
+                    self.error("expected a property name or array index")
 
             elif c == "]":
                 if len(self.path_stack) == 1:
@@ -1168,8 +1179,8 @@ class Lexer:
         return self.lex_inside_liquid_tag
 
 
-def tokenize(source: str) -> list[TokenT]:
+def tokenize(env: Environment, source: str) -> list[TokenT]:
     """Scan Liquid template _source_ and return a list of Markup objects."""
-    lexer = Lexer(source)
+    lexer = Lexer(env, source)
     lexer.run()
     return lexer.markup