+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Python Liquid2 Change Log
 
+## Version 0.4.0 (unreleased)
+
+**Features**
+
+- Added the `shorthand_indexes` class variable to `liquid2.Environment`. When `shorthand_indexes` is set to `True` (the default is `False`), array indexes in variable paths need not be surrounded by square brackets.
+
+**Changes**
+
+- `liquid2.tokenize` and `liquid2.lexer.Lexer` now require the current `Environment` to be passed as the first argument.
+
 ## Version 0.3.0
 
 **Breaking changes**